Move log configs to dedicated package.
